fixed controller, filter

+++ controllers/contacts.js
@@ -10,18 +10,27 @@ const {
 const getAll = async (req, res) => {
   const { _id: owner } = req.user;
   const { favorite = [true, false], page = 1, limit = 10 } = req.query;
+
+  if (favorite) {
+    const contacts = await Contact.find({
+      owner,
+      favorite: favorite,
+    });
+    res.json({ contacts });
+  }
+
   const skip = (page - 1) * limit;
-  const queryParams = { owner, favorite: favorite };
-  const result = await Contact.find(queryParams, "-createdAt -updatedAt", {
+  const result = await Contact.find({ owner }, "-createdAt -updatedAt", {
     skip,
     limit,
-  }).populate("owner", " email");
-  res.json(result);
+  }).populate("owner", "email");
+  res.json({ result });
 };
 
 const getById = async (req, res) => {
-  const { id } = req.params;
-  const result = await Contact.findOne({ _id: id });
+  const { _id: owner } = req.user;
+  const { contactId } = req.params;
+  const result = await Contact.findOne({ _id: contactId, owner: owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -39,13 +48,20 @@ const addContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
+  const { _id: owner } = req.user;
+
   const { error } = updateContactSchema.validate(req.body);
   if (error) {
     throw HttpError(400, "missing fields");
   }
 
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const result = await Contact.findByIdAndUpdate(
+    id,
+    req.body,
+    { owner: owner },
+    { new: true }
+  );
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -59,7 +75,11 @@ const updateStatusContact = async (req, res) => {
   }
 
   const { id } = req.params;
-  const result = await Contact.findByIdAndUpdate(id, req.body, { new: true });
+  const { _id: owner } = req.user;
+  const result = await Contact.findByIdAndUpdate(id, req.body, {
+    new: true,
+    owner: owner,
+  });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -67,8 +87,9 @@ const updateStatusContact = async (req, res) => {
 };
 
 const deleteContact = async (req, res) => {
+  const { _id: owner } = req.user;
   const { id } = req.params;
-  const result = await Contact.findByIdAndRemove(id);
+  const result = await Contact.findByIdAndRemove(id, { owner: owner });
   if (!result) {
     throw HttpError(404, "not found");
   }
